AmazonHomePage class added
