+++ src/test/java/pages/HomePage.java
@@ -1,0 +1,11 @@
+package pages;
+
+import org.openqa.selenium.support.PageFactory;
+import utilities.Driver;
+
+public class HomePage {
+
+    public HomePage(){
+        PageFactory.initElements(Driver.getDriver(), this);
+    }
+}
